feat(data): organize data, slight renaming, add new skills file
