+++ src/data/skillcolors.ts
