+++ src/data/skills.ts
@@ -1,0 +1,16 @@
+export default [
+  { name: 'Agility', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Construction', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Cooking', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Crafting', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Farming', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Firemaking', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Fishing', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Fletching', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Herblore', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Hunter', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Mining', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Runecrafting', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Smithing', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
+  { name: 'Woodcutting', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 }
+]
