+++ src/store.ts
@@ -1,33 +1,19 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
-import skillData from '@/skilldata';
-import leveltable from './leveltable';
-import skillColors from './skillcolors';
+import levelTable from './data/leveltable';
+import skills from './data/skills'
+import skillData from './data/skilldata';
+import skillColors from './data/skillcolors';
 
 Vue.use(Vuex);
 
 export default new Vuex.Store({
     state: {
         activeSkill: null,
-        skills: [
-            { name: 'Agility', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Construction', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Cooking', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Crafting', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Farming', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Firemaking', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Fishing', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Fletching', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Herblore', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Hunter', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Mining', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Runecrafting', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Smithing', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 },
-            { name: 'Woodcutting', playerLevel: 0, playerXp: 0, targetLevel: 0, targetXp: 0 }
-        ],
+        skills,
         skillColors,
         skillData,
-        leveltable
+        levelTable
     },
 
     mutations: {
@@ -50,15 +36,15 @@ export default new Vuex.Store({
         updateLevelFromXp(state, payload) {
             let i = 0;
             let row = 0;
-            for (let entry of leveltable) {
+            for (let entry of levelTable) {
                 if (entry.xp > payload.updatedSkill[payload.field]) { row = entry.level - 2; break; }
             }
             for (let skill of state.skills) {
                 if (skill.name === state.activeSkill) {
                     if (payload.field.includes('player')) {
-                        state.skills[i].playerLevel = leveltable[row].level;
+                        state.skills[i].playerLevel = levelTable[row].level;
                     } else {
-                        state.skills[i].targetLevel = leveltable[row].level;
+                        state.skills[i].targetLevel = levelTable[row].level;
                     }
                     break;
                 }
@@ -68,15 +54,15 @@ export default new Vuex.Store({
         updateXpFromLevel(state, payload) {
             let i = 0;
             let row = 0;
-            for (let entry of leveltable) {
+            for (let entry of levelTable) {
                 if (entry.level == payload.updatedSkill[payload.field]) { row = entry.level - 1; break; }
             }
             for (let skill of state.skills) {
                 if (skill.name === state.activeSkill) {
                     if (payload.field.includes('player')) {
-                        state.skills[i].playerXp = leveltable[row].xp;
+                        state.skills[i].playerXp = levelTable[row].xp;
                     } else {
-                        state.skills[i].targetXp = leveltable[row].xp;
+                        state.skills[i].targetXp = levelTable[row].xp;
                     }
                     break;
                 }
